feat: Refactor search modules by consolidating search logic into a controller and implementing specific search classes for Elaw, Esaj, and Projudi
